feat: Refactor forecast plugin to simple class architecture
- Replace inheritance-based ForecastPlugin with simple ForecastAnalyzerJob class
- Add plugin-wrapper.ts for backward compatibility with PluginManager
- Improve location finding to use chunk positions instead of searching entire document
- Clarify terminology: 'forecasting claims' (from document) vs 'our forecasts' (our estimates)
- Update variable names for clarity: forecastingClaims, ourForecasts
- Add comprehensive documentation (README, COMPARISON, MIGRATION_NOTES, etc.)
- Fix test compatibility issues with new architecture
- Leverage all data from extract-forecasting-claims tool (quality scores, rewritten text)

The new implementation is simpler, more accurate, and easier to maintain while
preserving full backward compatibility through the wrapper.
